requirements.txt
Faker==40.15.0
kafka-python-ng==2.2.3
black==26.3.1
flake8==7.3.0
pydantic==2.13.3
email-validator==2.2.0
